커뮤니티 완료

+++ src/main/webapp/WEB-INF/views/board_community/debate/debateViewMemo.jsp
@@ -12,12 +12,34 @@
 <html>
 <head>
 <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">  
+<link href="https://fonts.googleapis.com/css?family=Nanum+Gothic|Noto+Sans+KR&display=swap" rel="stylesheet">
 <link href="http://fonts.googleapis.com/css?family=Source+Sans+Pro:200,300,400,600,700,900" rel="stylesheet" />
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/default.css" >
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/fonts.css" >
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/body.css" >
 <title>토론방게시판 글보기</title>
 </head>
+
+<!-- *********************** 게시판 글쓰기 폼 ****************************  -->	
+<jsp:include page="../../header.jsp"></jsp:include>
+<!-- *********************** 사이드 메뉴 ****************************  -->	
+	
+<div id="body_div">
+	<div id="side_menu">
+		<h4><a href="/app/reviewList.do?current_page=1">커뮤니티</a></h4>
+		<div id="side_div">
+			<ul id="side_submenu">
+				<li>- <a href="/app/reviewList.do?current_page=1"> 리 뷰</a></li>
+				<li>- <a href="/app/freeList.do?current_page=1"> 자유게시판</a></li>
+				<li>- <a href="/app/meetingList.do?current_page=1"> 모임방</a></li>
+				<li>- <a href="/app/debateList.do?current_page=1"> 토론방</a></li>
+				<li>- <a href="/app/investList.do?current_page=1"> 재테크노하우</a></li>
+				<li>- <a href="/app/bestList.do?current_page=1"> BEST게시물</a></li>
+			</ul>
+		</div>
+	</div>
+	
+<!-- *********************** 내용 ****************************  -->	
 
 <!-- model에 저장해 둔거 불러옴 -->
 <c:set var="idx_num" value="${idx_num}" />
@@ -36,7 +58,7 @@ function boardlist(){
 	if(s=="None")
 		location.href='debateList.do?current_page=${current_page}';
 	else
-		location.href='debatesearch.do?pageForView=${current_page}&searchStr=${searchStr}';	
+		location.href='investSearchedList.do.do?pageForView=${current_page}&searchStr=${searchStr}';	
 }
 
 function boardmodify(){
@@ -44,7 +66,11 @@ function boardmodify(){
 }
 
 function boarddelete(){
-	location.href='debate_DeleteSpecificRow.do?idx_num=${idx_num}&current_page=${current_page}';	
+	if(confirm('정말 삭제하시겠습니까?')==true){
+		location.href='debate_DeleteSpecificRow.do?idx_num=${idx_num}&current_page=${current_page}';
+	}else {
+		return;
+	}
 }
 
 function boardrewrite(){
@@ -57,70 +83,45 @@ function download(){
 }
 </script>
 
-<!-- *********************** 게시판 글쓰기 폼 ****************************  -->	
-<jsp:include page="../../header.jsp"></jsp:include>
-<!-- *********************** 사이드 메뉴 ****************************  -->	
-	
-<div id="body_div">
-	<div id="side_menu">
-		<h4><a href="/app/reviewList.do?current_page=1">커뮤니티</a></h4>
-		<div id="side_div">
-			<ul id="side_submenu">
-				<li>- <a href="/app/reviewList.do?current_page=1"> 리 뷰</a></li>
-				<li>- <a href="/app/freeList.do?current_page=1"> 자유게시판</a></li>
-				<li>- <a href="/app/meetingList.do?current_page=1"> 모임방</a></li>
-				<li>- <a href="/app/debateList.do?current_page=1"> 토론방</a></li>
-				<li>- <a href="/app/investList.do?current_page=1"> 제태크노하우</a></li>
-				<li>- <a href="/app/bestList.do?current_page=1"> BEST게시물</a></li>
-			</ul>
-		</div>
-	</div>
-	
-<!-- *********************** 내용 ****************************  -->	
-
 <div id="line_div">
  <div id="sub_logo">
 <h2>토론방게시판</h2>	
- </div> 
+    </div> 
  <div id="site_div">	
-
-<input type="hidden" value="${boardData.getFilename()}">
-<table cellspacing="0" cellpadding="5" border="1" width="500">
-	
+<table  class="tlb_board">
 	<tr>
-		<td><b>조회수</b></td><td><c:out value="${boardData.getHits()}"/></td>
+			<td style="background-color: #f2f2f2"><b>제목</b></td>
+			<td><c:out value="${boardData.getSubject()}" /></td>
+			<td style="background-color: #f2f2f2"><b>조회수</b></td>
+			<td><c:out value="${boardData.getHits()}" /></td>
 	</tr>
 	<tr>
-		<td><b>작성일</b></td><td><c:out value="${boardData.getCreated_date()}"/></td>
+			<td style="background-color: #f2f2f2"><b>아이디</b></td>
+			<td><c:out value="${boardData.getId()}" /></td>
+			<td style="background-color: #f2f2f2"><b>작성일</b></td>
+			<td><c:out value="${boardData.getCreated_date()}"></c:out></td>
 	</tr>
 	<tr>
-		<td><b>아이디</b></td><td><c:out value="${boardData.getId()}"/></td>
+		<td style="background-color: #f2f2f2"><b>내용</b></td>
+		<td  colspan="3" height="300" style="text-align: left; padding-left: 15px;">
+		<c:out value="${boardData.getContent()}"/></td>
+		</td>
 	</tr>
 	<tr>
-		<td><b>제목</b></td><td><c:out value="${boardData.getSubject()}"/></td>
-	</tr>
-	<tr>
-		<td><b>내용</b></td><td><c:out value="${boardData.getContent()}"/></td>
-	</tr>
-	<tr>
-		<td>파일첨부</td>
-		<td><input type="button"
-		onclick="download()" 
-		value="${boardData.getFilename()}"/></td>
-	</tr>
+			<td style="background-color: #f2f2f2"><b>파일첨부</b></td>
+			<td colspan="3" style="text-align: left; padding-left: 15px; ">
+			<input type="button" onclick="download()"
+			value="${boardData.getFilename()}">
+			</td>
+		</tr>
 </table>
 
-<table cellspacing="0" cellpadding="0" border="0" width="500">
-	<tr>
-	<td>
-		<input type="button" value="수정" onclick="javascript:boardmodify()">
-		<input type="button" value="목록" onclick="javascript:boardlist()">
-		<input type="button" value="답글" onclick="javascript:boardrewrite()">
-		<input type="button" value="삭제" onclick="javascript:boarddelete()">
-	</td>
-	</tr>
-
-</table>
+<div class="div_board_bnt">
+		<input type="button" value="수 정" onclick="javascript:boardmodify()" class="bnt_view">
+		<input type="button" value="목 록" onclick="javascript:boardlist()" class="bnt_view">
+		<input type="button" value="답 글" onclick="javascript:boardrewrite()" class="bnt_view">
+		<input type="button" value="삭 제" onclick="javascript:boarddelete()" class="bnt_view">
+	</div>
 </div>
 </div>
  </div>
