+++ src/main/webapp/WEB-INF/views/board_community/invest/investViewMemo.jsp
@@ -1,25 +1,31 @@
+<%@page import="java.nio.file.Path"%>
+<%@page import="org.springframework.core.io.ClassPathResource"%>
+<%@page import="com.ourbank.app.bean.InvestBoard_Bean"%>
 <%@ page language="java"
 	import="java.util.*,java.sql.*,javax.servlet.http.*"
 	contentType="text/html; charset=EUC-KR" pageEncoding="EUC-KR"%>
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" 
 "http://www.w3.org/TR/html4/loose.dtd">
 <%@ page import="java.io.*,java.text.*"%>
-
 <%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
-<!-- model에 저장해 둔거 불러옴 -->
 
-<c:set var="str_aid" value="${idx }" />
-<c:set var="str_c_page" value="${current_page}" />
-<c:set var="searchString" value="${searchStr}" />
-<c:set var="filename" value="${filename}"/>
 <html>
 <head>
 <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">  
+<link href="https://fonts.googleapis.com/css?family=Nanum+Gothic|Noto+Sans+KR&display=swap" rel="stylesheet">
 <link href="http://fonts.googleapis.com/css?family=Source+Sans+Pro:200,300,400,600,700,900" rel="stylesheet" />
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/default.css" >
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/fonts.css" >
 <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/body.css" >
-<title>invest</title>
+<title>재테크노하우  글보기</title>
+</head>
+
+<!-- model에 저장해 둔거 불러옴 -->
+<c:set var="str_aid" value="${idx }" />
+<c:set var="str_c_page" value="${current_page}" />
+<c:set var="searchString" value="${searchStr}" />
+<c:set var="filename" value="${filename}"/>
+
 <%
 	String searchString=(String) (pageContext.getAttribute("searchString"));
 	String filename=(String) (pageContext.getAttribute("filename"));
@@ -30,7 +36,7 @@ function boardlist(){
 	if(s=="None")
 		location.href='investList.do?current_page=${current_page}';
 	else
-		location.href='listSearchedSpecificPageWork.do?pageForView=${current_page}&searchStr=${searchStr}';	
+		location.href='investSearchedList.do?pageForView=${current_page}&searchStr=${searchStr}';	
 }
 
 function boardmodify(){
@@ -38,7 +44,11 @@ function boardmodify(){
 }
 
 function boarddelete(){
-	location.href='investDeleteSpecificRow.do?idx=${idx}&current_page=${current_page}';	
+	if(confirm('정말 삭제하시겠습니까?')==true){
+		location.href='investDeleteSpecificRow.do?idx=${idx}&current_page=${current_page}'
+	}else {
+		return;
+	}
 }
 
 function boardrewrite(){
@@ -52,7 +62,6 @@ function download(){
 }
 
 </script>
-</head>
 <body>
 <!-- *********************** 게시판 글쓰기 폼 ****************************  -->	
 	<jsp:include page="../../header.jsp"></jsp:include>
@@ -68,7 +77,7 @@ function download(){
 				<li>- <a href="/app/freeList.do?current_page=1"> 자유게시판</a></li>
 				<li>- <a href="/app/meetingList.do?current_page=1"> 모임방</a></li>
 				<li>- <a href="/app/debateList.do?current_page=1"> 토론방</a></li>
-				<li>- <a href="/app/investList.do?current_page=1"> 제태크노하우</a></li>
+				<li>- <a href="/app/investList.do?current_page=1"> 재테크노하우</a></li>
 				<li>- <a href="/app/bestList.do?current_page=1"> BEST게시물</a></li>
 			</ul>
 		</div>
@@ -78,58 +87,53 @@ function download(){
 
 <div id="line_div">
  <div id="sub_logo">
- <h2>제태크 노하우</h2>
+ <h2>재테크 노하우</h2>
  </div> 
- <div id="site_div">
-	
-<table cellspacing="0" cellpadding="5" border="1" width="600"  >
+ <div id="site_div">	
+<table  class="tlb_board">
 		<tr>
-			<td><b>제목</b></td>
+			<td style="background-color: #f2f2f2"><b>제목</b></td>
 			<td><c:out value="${boardData.getSubject()}" /></td>
-			<td><b>조회수</b></td>
+			<td style="background-color: #f2f2f2"><b>조회수</b></td>
 			<td><c:out value="${boardData.getHits()}" /></td>
 		</tr>
 		<tr>
-			<td><b>ID</b></td>
+			<td style="background-color: #f2f2f2"><b>아이디</b></td>
 			<td><c:out value="${boardData.getId()}" /></td>
-			<td><b>작성일</b></td>
+			<td style="background-color: #f2f2f2"><b>작성일</b></td>
 			<td><c:out value="${boardData.getCreated_date()}"></c:out></td>
 		</tr>
 		<tr>
-			<td><b>내용</b></td>
-			<td  colspan="3" height="400" >
+			<td style="background-color: #f2f2f2"><b>내용</b></td>
+			<td  colspan="3" height="300" style="text-align: left; padding-left: 15px;">
 			<c:if test="${boardData.getDepth()==0}">
 			[<c:out value="${boardData.getInvest_case()}"/>]<br><br>
 			</c:if>
 			<c:out value="${boardData.getContent()}" /></td>
 		</tr>
 		<tr>
-			<td><b>파일첨부</b></td>
-			<td colspan="3">
+			<td style="background-color: #f2f2f2"><b>파일첨부</b></td>
+			<td colspan="3" style="text-align: left; padding-left: 15px; ">
 			<input type="button" onclick="download()"
 			value="${boardData.getFilename()}">
 			</td>
 		</tr>
 	</table>
 
-<table cellspacing="0" cellpadding="0" border="0" width="500">
-	<tr>
-	<td>
-		<input type="button" value="수정" onclick="javascript:boardmodify()">
-		<input type="button" value="목록" onclick="javascript:boardlist()">
-		<input type="button" value="답글" onclick="javascript:boardrewrite()">
-		<input type="button" value="삭제" onclick="javascript:boarddelete()">
-	</td>
-	</tr>
+	<div class="div_board_bnt">
+		<input type="button" value="수 정" onclick="javascript:boardmodify()" class="bnt_view">
+		<input type="button" value="목 록" onclick="javascript:boardlist()" class="bnt_view">
+		<input type="button" value="답 글" onclick="javascript:boardrewrite()" class="bnt_view">
+		<input type="button" value="삭 제" onclick="javascript:boarddelete()" class="bnt_view">
+	</div>
 
-</table>
 </div>
 </div>
-</div>
+</div>	
+
 
 <!-- *********************** 게시판 글쓰기 폼 ****************************  -->	
 
 	<jsp:include page="../../footer.jsp"></jsp:include>
 	
 </body>
-</html>
